Fix bug with creating and deleting

+++ packages/server/src/api/services/publish/updates/delete.ts
@@ -1,6 +1,7 @@
 import { parseUpdate } from '@harmony/util/src/updates/utils'
 import { deleteComponentSchema } from '@harmony/util/src/updates/component'
 import type { ComponentUpdate } from '@harmony/util/src/types/component'
+import { getLevelId } from '@harmony/util/src/utils/component'
 import type { FlowGraph } from '../../indexor/graph'
 import type { UpdateComponent } from './types'
 
@@ -11,14 +12,24 @@ export const deleteUpdate: UpdateComponent = ({ value, update }, graph) => {
 }
 
 export const deleteComponent = (update: ComponentUpdate, graph: FlowGraph) => {
-  const deleteElement = graph.getJSXElementById(
-    update.componentId,
-    update.childIndex,
-  )
-  if (!deleteElement) {
-    throw new Error(
-      `Element with id ${update.componentId} and ${update.childIndex} not found`,
+  const numLevels = update.componentId.split('#').length
+  for (let i = 0; i < numLevels; i++) {
+    const componentId = getLevelId(update.componentId, i)
+    const deleteElement = graph.getJSXElementById(
+      componentId,
+      update.childIndex,
     )
+    if (!deleteElement) {
+      throw new Error(
+        `Element with id ${update.componentId} and ${update.childIndex} not found`,
+      )
+    }
+    const parentComponent = deleteElement.getParentComponent()
+    if (parentComponent.getJSXElements()[0].id === deleteElement.id) {
+      continue
+    }
+    return graph.deleteElement(deleteElement)
   }
-  return graph.deleteElement(deleteElement)
+
+  throw new Error('Could not find element to delete')
 }
